+++ docs/conf.py
@@ -32,6 +32,7 @@ release = '0.10.0'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
+    'sphinx.ext.autosectionlabel',
     'sphinx.ext.autodoc',
     'sphinx_rtd_theme'
 ]